fix: revert get_avatar to skill-scoped tool and improve skill triggers
Making get_avatar always-loaded adds token overhead to every conversation.
Reverted to skill-scoped tool in settings skill and improved activation:

- Restored get_avatar in settings TOOLS.json and bundled-tool-registry
- Removed global avatar-read.ts from tool-manifest
- Enhanced settings skill description with "view, show" keywords
- Added get_avatar usage docs to settings SKILL.md
- Reverted guard test back to 11 tools
