+++ assistant/src/__tests__/always-loaded-tools-guard.test.ts
@@ -28,7 +28,7 @@ afterAll(() => {
 });
 
 describe("always-loaded tool count", () => {
-  test("should be exactly 12 (no-client baseline excludes host tools)", async () => {
+  test("should be exactly 11 (no-client baseline excludes host tools)", async () => {
     await initializeTools();
     const allDefs = buildToolDefinitions();
 
@@ -56,7 +56,6 @@ describe("always-loaded tool count", () => {
       "file_edit",
       "file_read",
       "file_write",
-      "get_avatar",
       "memory_manage",
       "memory_recall",
       "skill_execute",
@@ -66,6 +65,6 @@ describe("always-loaded tool count", () => {
     ].sort();
 
     expect(activeNames).toEqual(expectedNames);
-    expect(activeTools.length).toBe(12);
+    expect(activeTools.length).toBe(11);
   });
 });
