+++ assistant/src/config/bundled-tool-registry.ts
@@ -132,6 +132,7 @@ import * as sequenceImport from "./bundled-skills/sequences/tools/sequence-impor
 import * as sequenceList from "./bundled-skills/sequences/tools/sequence-list.js";
 import * as sequenceUpdate from "./bundled-skills/sequences/tools/sequence-update.js";
 // ── settings ───────────────────────────────────────────────────────────────────
+import * as avatarGet from "./bundled-skills/settings/tools/avatar-get.js";
 import * as avatarRemove from "./bundled-skills/settings/tools/avatar-remove.js";
 import * as avatarUpdate from "./bundled-skills/settings/tools/avatar-update.js";
 import * as navigateSettingsTab from "./bundled-skills/settings/tools/navigate-settings-tab.js";
@@ -316,6 +317,7 @@ export const bundledToolRegistry = new Map<string, SkillToolScript>([
   ["sequences:tools/sequence-analytics.ts", sequenceAnalytics],
 
   // settings
+  ["settings:tools/avatar-get.ts", avatarGet],
   ["settings:tools/avatar-update.ts", avatarUpdate],
   ["settings:tools/avatar-remove.ts", avatarRemove],
   ["settings:tools/voice-config-update.ts", voiceConfigUpdate],
